Fix de plantillas Fix de relaciones Eloquent: - Localidades, - Cargos, - Turnos, - Grupos, - Empleados, - Clientes, - Sucursales

+++ database/migrations/2020_08_14_220454_create_turnos_table.php
@@ -14,7 +14,7 @@ class CreateTurnosTable extends Migration
     public function up()
     {
         Schema::create('turnos', function (Blueprint $table) {
-            $table->tinyInteger('turno_empleado',true);
+            $table->smallInteger('turno_empleado',true);
             $table->string('descripcion')->nullable();;
             $table->time('hora_desde')->nullable();;
             $table->time('hora_hasta')->nullable();;
